+++ src/App.js
@@ -10,7 +10,7 @@ import {StyledBackground} from './components/UI';
 import {SideButton} from './components/UI/SideButton/SideButton';
 import {useConstants, useTracking} from './hooks';
 import {useApp} from './providers/AppProvider';
-import {Bridge, Liquidity, Faq, ProtectedRoute, Terms} from './routes';
+import {Bridge, Faq, ProtectedRoute, Terms} from './routes';
 import {openInNewTab} from './utils';
 
 export const App = () => {
