+++ src/components/Features/Source/Source.module.scss
@@ -1,0 +1,12 @@
+@import '../../../index';
+
+.source {
+  transition: 0.3s ease-in-out;
+
+  .tabsContainer {
+    display: flex;
+    background: $--color-alpha-1;
+    border-radius: 10px;
+    padding: 10px;
+  }
+}
