+++ src/components/Features/Transfer/Transfer.js
@@ -1,82 +1,199 @@
-import React from 'react';
+import {NetworkType} from '@starkware-industries/commons-js-enums';
+import PropTypes from 'prop-types';
+import React, {useEffect, useState} from 'react';
 
-import sourceMap, {SourceGroup} from '../../../config/sources';
-import {ActionType} from '../../../enums';
-import {useTransferTracking, useTransferTranslation} from '../../../hooks';
-import {useSource} from '../../../providers/SourceProvider';
-import {useIsL1, useIsL2, useTransfer} from '../../../providers/TransferProvider';
-import {Menu, TransferMenuTab} from '../../UI';
-import {EthereumBridge} from '../EthereumBridge/EthereumBridge';
-import {OtherBridge} from '../OtherBridge/OtherBridge';
+import {
+  useTransferToL1,
+  useTransferToL2,
+  useTransferTracking,
+  useTransferTranslation
+} from '../../../hooks';
+import {useLogin} from '../../../providers/AppProvider';
+import {useMenu} from '../../../providers/MenuProvider';
+import {useL1Token, useL2Token, useTokens} from '../../../providers/TokensProvider';
+import {
+  useAmount,
+  useIsL1,
+  useIsL2,
+  useBridgeIsFull,
+  useTransfer
+} from '../../../providers/TransferProvider';
+import {afterDecimal, evaluate, isNegative, isZero} from '../../../utils';
+import {
+  Loading,
+  LoadingSize,
+  NetworkMenu,
+  NetworkSwap,
+  TokenInput,
+  TransferButton,
+  LoginWalletButton,
+  MenuBackground
+} from '../../UI';
 import styles from './Transfer.module.scss';
 
-export const Transfer = () => {
-  const [trackSwapNetworks] = useTransferTracking();
-  const [, swapToL1] = useIsL1();
-  const [isL2, swapToL2] = useIsL2();
-  const {action} = useTransfer();
-  const {depositTxt, withdrawTxt} = useTransferTranslation();
-  const {group, selectGroup, selectSource} = useSource();
+export const Transfer = ({onNetworkSwap}) => {
+  const {
+    insufficientBalanceErrorMsg,
+    maxDepositErrorMsg,
+    tooManyDigitsErrorMsg,
+    negativeValueErrorMsg
+  } = useTransferTranslation();
+  const [trackMaxClick] = useTransferTracking();
+  const [isL1] = useIsL1();
+  const [isL2] = useIsL2();
+  const [amount, setAmount] = useAmount();
+  const [hasInputError, setHasInputError] = useState(false);
+  const [errorMsg, setErrorMsg] = useState('');
+  const [isButtonDisabled, setIsButtonDisabled] = useState(true);
+  const {showSelectTokenMenu} = useMenu();
+  const {selectToken, selectedToken} = useTransfer();
+  const {tokens, updateTokenBalance} = useTokens();
+  const {bridgeIsFull} = useBridgeIsFull();
+  const transferToL2 = useTransferToL2();
+  const transferToL1 = useTransferToL1();
+  const getL1Token = useL1Token();
+  const getL2Token = useL2Token();
+  const {isLoggedIn} = useLogin();
 
-  const tabs = [
-    {
-      text: depositTxt,
-      isActive: action === ActionType.TRANSFER_TO_L2,
-      onClick: () => {
-        if (!sourceMap[group].deposit_sources && !sourceMap[group].sources) {
-          selectGroup(SourceGroup.ETHEREUM);
-          selectSource(sourceMap[SourceGroup.ETHEREUM].sources[0]);
-        }
-        onNetworkTabClick(ActionType.TRANSFER_TO_L2);
-      }
-    },
-    {
-      text: withdrawTxt,
-      isActive: action === ActionType.TRANSFER_TO_L1,
-      onClick: () => {
-        if (!sourceMap[group].withdraw_sources && !sourceMap[group].sources) {
-          selectGroup(SourceGroup.ETHEREUM);
-          selectSource(sourceMap[SourceGroup.ETHEREUM].sources[0]);
-        }
-        onNetworkTabClick(ActionType.TRANSFER_TO_L1);
+  useEffect(() => {
+    if (!selectedToken) {
+      selectToken(tokens[0].symbol);
+    }
+  }, []);
+
+  useEffect(() => {
+    if (selectedToken) {
+      const {isLoading, maxDeposit} = selectedToken;
+      setHasInputError(false);
+      if (isLoading || isZero(amount) || (isL1 && !maxDeposit)) {
+        setIsButtonDisabled(true);
+      } else {
+        validateAmount();
       }
     }
-  ];
+  }, [amount, selectedToken, isL1]);
 
-  const onSwapClick = () => {
-    trackSwapNetworks();
-    isL2 ? swapToL1() : swapToL2();
-  };
+  const validateAmount = () => {
+    let errorMsg = '';
+    const {decimals, balance, maxDeposit, symbol} = selectedToken;
 
-  const onNetworkTabClick = tab => {
-    if (action !== tab) {
-      onSwapClick();
+    if (afterDecimal(amount) > decimals) {
+      errorMsg = tooManyDigitsErrorMsg;
+    } else if (isNegative(amount)) {
+      errorMsg = negativeValueErrorMsg;
+    } else if (Number(amount) > Number(balance)) {
+      errorMsg = insufficientBalanceErrorMsg;
+    } else if (isL1 && Number(amount) > Number(maxDeposit)) {
+      errorMsg = evaluate(maxDepositErrorMsg, {maxDeposit, symbol});
+    }
+
+    if (errorMsg) {
+      setHasInputError(true);
+      setErrorMsg(errorMsg);
+      setIsButtonDisabled(true);
+    } else {
+      setIsButtonDisabled(false);
     }
   };
 
-  const renderTabs = () => {
-    return tabs.map((tab, index) => {
-      return (
-        <TransferMenuTab
-          key={index}
-          isActive={tab.isActive}
-          text={tab.text}
-          onClick={tab.onClick}
+  const onMaxClick = () => {
+    const {balance, maxDeposit} = selectedToken;
+    try {
+      trackMaxClick();
+      setAmount(String(Math.min(balance, isL1 ? Number(maxDeposit) : Infinity)));
+    } catch (ex) {
+      setAmount(balance);
+    }
+  };
+
+  const onInputChange = event => {
+    setAmount(event.target.value);
+  };
+
+  const onTransferClick = async () => (isL1 ? transferToL2(amount) : transferToL1(amount));
+
+  const onRefreshTokenBalanceClick = () => {
+    updateTokenBalance(selectedToken.symbol);
+  };
+
+  const renderL1Network = () => {
+    const tokenData = getL1Token(selectedToken.symbol);
+    return (
+      <NetworkMenu
+        isDisabled={bridgeIsFull}
+        isTarget={!isL1}
+        networkName={NetworkType.L1}
+        tokenData={tokenData}
+        onRefreshClick={onRefreshTokenBalanceClick}
+      >
+        {isL1 && renderTransferInput()}
+      </NetworkMenu>
+    );
+  };
+
+  const renderL2Network = () => {
+    const tokenData = getL2Token(selectedToken.symbol);
+    return (
+      <NetworkMenu
+        isDisabled={bridgeIsFull}
+        isTarget={!isL2}
+        networkName={NetworkType.L2}
+        tokenData={tokenData}
+        onRefreshClick={onRefreshTokenBalanceClick}
+      >
+        {isL2 && renderTransferInput()}
+      </NetworkMenu>
+    );
+  };
+
+  const renderTransferInput = () => {
+    return (
+      <>
+        <TokenInput
+          hasError={hasInputError}
+          isInputDisabled={bridgeIsFull}
+          tokenData={selectedToken}
+          value={amount}
+          onInputChange={onInputChange}
+          onMaxClick={onMaxClick}
+          onTokenSelect={showSelectTokenMenu}
         />
-      );
-    });
+        {hasInputError && <div className={styles.errorMsg}>{errorMsg}</div>}
+      </>
+    );
   };
 
   return (
-    <Menu>
-      <div className={styles.transfer}>
-        <div className={styles.tabsContainer}>{renderTabs()}</div>
-        {group === SourceGroup.ETHEREUM ? (
-          <EthereumBridge onSwapClick={onSwapClick} />
-        ) : (
-          <OtherBridge />
-        )}
-      </div>
-    </Menu>
+    <MenuBackground>
+      {!selectedToken && (
+        <center>
+          <Loading size={LoadingSize.XL} />
+        </center>
+      )}
+      {selectedToken && (
+        <>
+          <div className={styles.fromMenu}>{isL1 ? renderL1Network() : renderL2Network()}</div>
+          <NetworkSwap isFlipped={isL2} onClick={onNetworkSwap} />
+          <div className={styles.divider} />
+          <div className={styles.toMenu}>{isL1 ? renderL2Network() : renderL1Network()}</div>
+          <div className={styles.divider} />
+
+          <div className={styles.transferButtonContainer}>
+            {isLoggedIn ? (
+              <TransferButton
+                isDisabled={isButtonDisabled || bridgeIsFull}
+                onClick={onTransferClick}
+              />
+            ) : (
+              <LoginWalletButton />
+            )}
+          </div>
+        </>
+      )}
+    </MenuBackground>
   );
+};
+
+Transfer.propTypes = {
+  onNetworkSwap: PropTypes.func
 };
