+++ src/components/Features/index.js
@@ -1,5 +1,5 @@
 export * from './Account/Account';
 export * from './SelectToken/SelectToken';
-export * from './Transfer/Transfer';
+export * from './Source/Source';
 export * from './TransferLog/TransferLog';
 export * from './ToastManager/ToastManager';
