--- src/components/UI/LiquidityButton/LiquidityButton.module.scss
@@ -1,3 +1,0 @@
-.liquidityButton {
-  border-width: 2px;
-}
