+++ src/components/UI/MultiChoiceList/MultiChoiceList.module.scss
@@ -2,7 +2,6 @@
 
 .multiChoiceList {
   .container {
-    margin-top: 20px;
     border: 1px solid $--color-alpha-3;
     box-sizing: border-box;
     border-radius: 8px;
