+++ src/components/UI/TokenInput/TokenInput.module.scss
@@ -6,8 +6,8 @@
   align-items: center;
   background: $--color-alpha-2;
   height: 50px;
-  margin: 8px 0;
-  padding: 8px;
+  margin: 20px 0;
+  padding: 10px;
   border-radius: 7px;
   transition: 0.3s ease-in-out;
