+++ src/components/UI/index.js
@@ -48,5 +48,4 @@ export * from './LoginWalletButton/LoginWalletButton';
 export * from './NetworkWalletButton/NetworkWalletButton';
 export * from './MainMenuButton/MainMenuButton';
 export * from './ChainSelect/ChainSelect';
-export * from './LiquidityButton/LiquidityButton';
 export * from './MenuBackground/MenuBackground';
