+++ src/providers/SourceProvider/source-context.js
@@ -4,6 +4,6 @@ import {initialState} from './source-reducer';
 
 export const SourceContext = createContext({
   ...initialState,
-  selectGroup: source => ({source}),
-  selectSource: source => ({source})
+  selectSource: source => ({source}),
+  selectDefaultSource: () => {}
 });
