--- src/routes/Liquidity/Liquidity.module.scss
@@ -1,3 +1,0 @@
-.liquidity {
-  min-width: 500px;
-}
