+++ src/routes/index.js
@@ -1,5 +1,4 @@
 export * from './Bridge/Bridge';
 export * from './Terms/Terms';
 export * from './Faq/Faq';
-export * from './Liquidity/Liquidity';
 export * from './ProtectedRoute';
